backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.34
pydantic-settings>=2.4.0
psycopg[binary]>=3.2.1
pandas>=2.2.2
openpyxl>=3.1.5
python-dotenv>=1.0.1
httpx>=0.28.1
xlrd>=2.0.1
